feat: stamp CanTpConnection, CanTpEcu, CanTpNode, CanNmEcu (R23-11); move AbstractCanPhysicalChannel and CanPhysicalChannel to Fibex4Can/CanTopology
